Moved images and css in matching component
