+++ frontend/src/components/Matching/IconMatching.tsx
@@ -5,8 +5,8 @@ function importAll(r: any) {
   return r.keys().map(r);
 }
 
-const iconContext = require.context('./icons', false, /\.png$/);
-const activeIconContext = require.context('./icons/active', false, /\.png$/);
+const iconContext = require.context('../../images/icons', false, /\.png$/);
+const activeIconContext = require.context('../../images/icons/active', false, /\.png$/);
 
 const icons: any[] = importAll(iconContext);
 const activeIcons: any[] = importAll(activeIconContext);
