+++ frontend/src/images/icons/active/array_active.png
