+++ frontend/src/images/icons/active/binary_search_active.png
