+++ frontend/src/images/icons/active/binary_search_tree_active.png
