+++ frontend/src/images/icons/active/bit_manipulation_active.png
