+++ frontend/src/images/icons/active/depth-_first_search_active.png
