+++ frontend/src/images/icons/active/greedy_active.png
