+++ frontend/src/images/icons/active/hash_table_active.png
