+++ frontend/src/images/icons/active/simulation_active.png
