+++ frontend/src/images/icons/active/string_active.png
