+++ frontend/src/images/icons/active/trie_active.png
