+++ frontend/src/images/icons/active/two_pointers_active.png
