+++ frontend/src/images/icons/binary_search.png
