+++ frontend/src/images/icons/binary_search_tree.png
