+++ frontend/src/images/icons/bit_manipulation.png
