+++ frontend/src/images/icons/depth-_first_search.png
