+++ frontend/src/images/icons/greedy.png
