+++ frontend/src/images/icons/hash_table.png
