+++ frontend/src/images/icons/simulation.png
