+++ frontend/src/images/icons/string.png
