+++ frontend/src/images/icons/trie.png
